move event_name_to_contract to forkable state
